+++ examples/example3d.rs
@@ -126,11 +126,14 @@ fn setup_arena(
     mut mesh_assets: ResMut<Assets<Mesh>>,
     mut material_assets: ResMut<Assets<StandardMaterial>>,
 ) {
-    let mesh = mesh_assets.add(Mesh::from(shape::Plane {
-        size: 100.0,
-        subdivisions: 0,
-    }));
-    let material = material_assets.add(Color::GRAY.into());
+    let mesh = mesh_assets.add(Mesh::from(
+        Plane3d {
+            normal: Direction3d::Y,
+        }
+        .mesh()
+        .size(100.0, 100.0),
+    ));
+    let material = material_assets.add(Color::GRAY);
     commands.spawn(PbrBundle {
         mesh,
         material,
@@ -180,15 +183,15 @@ fn populate_planet(
 fn control_spaceship(
     mut spaceship_query: Query<&mut Transform, With<IsSpaceship>>,
     time: Res<Time>,
-    input: Res<Input<KeyCode>>,
+    input: Res<ButtonInput<KeyCode>>,
 ) {
     let calc_axis = |neg: KeyCode, pos: KeyCode| match (input.pressed(neg), input.pressed(pos)) {
         (true, true) | (false, false) => 0.0,
         (true, false) => -1.0,
         (false, true) => 1.0,
     };
-    let pitch = calc_axis(KeyCode::Up, KeyCode::Down);
-    let roll = calc_axis(KeyCode::Left, KeyCode::Right);
+    let pitch = calc_axis(KeyCode::ArrowUp, KeyCode::ArrowDown);
+    let roll = calc_axis(KeyCode::ArrowLeft, KeyCode::ArrowRight);
     for mut spaceship_transform in spaceship_query.iter_mut() {
         let forward_direction = spaceship_transform.rotation.mul_vec3(-Vec3::Z);
         let roll_quat =
@@ -231,16 +234,10 @@ fn populate_simple_sphere(
     populate.populate(|ctx, mut cmd, ()| {
         if ctx.is_first_time() {
             let mesh = mesh
-                .get_or_insert_with(|| {
-                    mesh_assets.add(Mesh::from(shape::UVSphere {
-                        radius: 1.0,
-                        sectors: 10,
-                        stacks: 10,
-                    }))
-                })
+                .get_or_insert_with(|| mesh_assets.add(Mesh::from(Sphere { radius: 1.0 })))
                 .clone();
             let material = material
-                .get_or_insert_with(|| material_assets.add(Color::YELLOW.into()))
+                .get_or_insert_with(|| material_assets.add(Color::YELLOW))
                 .clone();
             cmd.insert(PbrBundle {
                 mesh,
@@ -275,7 +272,7 @@ fn edit_laser_pointer(
         };
         if button.clicked() {
             exclusive_queue.push_back(
-                vpeol_read_click_on_entity::<&YoleckEntityUuid>
+                vpeol_read_click_on_entity::<With<YoleckEntityUuid>>
                     .pipe(yoleck_map_entity_to_uuid)
                     .pipe(
                         |In(target): In<Option<Uuid>>, mut edit: YoleckEdit<&mut LaserPointer>| {